+++ libft.a/ft_strncmp.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strncmp.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tvan-gij </var/mail/tvan-gij>              +#+  +:+       +#+        */
+/*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/12/15 15:26:37 by tvan-gij          #+#    #+#             */
-/*   Updated: 2020/12/15 15:44:17 by tvan-gij         ###   ########.fr       */
+/*   Created: 2021/01/17 13:23:31 by tvan-gij          #+#    #+#             */
+/*   Updated: 2021/01/17 13:23:38 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strncmp(const char *s1, const char *s2, size_t n)
+int		ft_strncmp(const char *s1, const char *s2, unsigned int n)
 {
-	char	*str1;
-	char	*str2;
-	size_t	i;
+	char			*str1;
+	char			*str2;
+	unsigned int	i;
 
 	str1 = (char *)s1;
 	str2 = (char *)s2;
