+++ libft.a/ft_substr.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tvan-gij </var/mail/tvan-gij>              +#+  +:+       +#+        */
+/*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/12/18 13:24:13 by tvan-gij          #+#    #+#             */
-/*   Updated: 2020/12/18 13:49:40 by tvan-gij         ###   ########.fr       */
+/*   Created: 2021/01/17 13:26:05 by tvan-gij          #+#    #+#             */
+/*   Updated: 2021/01/17 13:26:11 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-char	*ft_substr(const char *s, unsigned int start, size_t len)
+char	*ft_substr(const char *s, unsigned int start, unsigned int len)
 {
-	char	*subs;
-	size_t	i;
+	char			*subs;
+	unsigned int	i;
 
 	i = start;
 	while (s[i])
